reference-implementation/requirements.txt
aniso8601==7.0.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
Rx==1.6.1
six==1.16.0
